Fix deployment such that the hosting works
Previous solution put `dao.json` which was obtained by hardhat deploy '--export-all' CLI flag.
This exported all deployments listed in 'hardhat/deployments' directory. Since this file was present
locally it worked. However, since it's ignored, it will not be present when building.

The solution is to create a placeholder deployment (checked in git), which will be copied to a localhost
deployment file, which is now gitignored. We also now import the supported chains one by one
(but this is not a big deal).